requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.1
Django==4.2.6
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
excel-base==1.0.4
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
isoweek==1.3.3
mysqlclient==2.2.0
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
PyJWT==1.7.1
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzlocal==5.1
uritemplate==4.1.1
xlwt==1.3.0
